Conflicts:
	English.lproj/DiffComputationBanner.xib
	English.lproj/FindReplaceBanner.xib
	English.lproj/JumpToOffsetBanner.xib
	English.lproj/MoveSelectionByBanner.xib
	English.lproj/SaveBanner.xib
	HexFiend_2.xcodeproj/project.pbxproj
	HexFiend_2.xcodeproj/xcshareddata/xcschemes/Debug + UnitTests.xcscheme
	app/sources/AppDelegate.m
	app/sources/BaseDataDocument.m
	app/sources/DataInspectorRepresenter.m
	app/sources/HFDocumentOperationView.h
	app/sources/HFDocumentOperationView.m
	framework/sources/HFByteSlice.m
	framework/sources/HFByteSliceFileOperation.m
	framework/sources/HFPasteboardOwner.m
	framework/sources/HFSharedMemoryByteSlice.m

+++ framework/sources/HFFileReference.m
@@ -275,7 +275,7 @@ static BOOL returnFTruncateError(NSError **error) {
 			tempBuf = malloc(blockSize);
 			ssize_t result = pread(fileDescriptor, tempBuf, blockSize, pos - prePad);
 			if (result != (ssize_t)blockSize) {
-				[NSException raise:NSGenericException format:@"Read result: %d expected: %u error: %s", result, blockSize, strerror(errno)];
+				[NSException raise:NSGenericException format:@"Read result: %zd expected: %u error: %s", result, blockSize, strerror(errno)];
 			}
 			NSUInteger toCopy = blockSize - prePad;
 			if (toCopy > length)
@@ -291,7 +291,7 @@ static BOOL returnFTruncateError(NSError **error) {
 
     ssize_t result = pread(fileDescriptor, buff, length, pos);
     if (result != (long)length) {
-        [NSException raise:NSGenericException format:@"Read result: %d expected: %u error: %s", result, length, strerror(errno)];
+        [NSException raise:NSGenericException format:@"Read result: %zd expected: %lu error: %s", result, (unsigned long)length, strerror(errno)];
     }
 
 	if (lastBlockLen) {
@@ -299,7 +299,7 @@ static BOOL returnFTruncateError(NSError **error) {
 			tempBuf = malloc(blockSize);
 		result = pread(fileDescriptor, tempBuf, blockSize, pos + length);
 		if (result != (ssize_t)blockSize) {
-			[NSException raise:NSGenericException format:@"Read result: %d expected: %u error: %s", result, blockSize, strerror(errno)];
+			[NSException raise:NSGenericException format:@"Read result: %zd expected: %u error: %s", result, blockSize, strerror(errno)];
 		}
 		memcpy(buff + length, tempBuf, lastBlockLen);
 	}
@@ -327,7 +327,7 @@ static BOOL returnFTruncateError(NSError **error) {
 			tempBuf = malloc(blockSize);
 			ssize_t result = pread(fileDescriptor, tempBuf, blockSize, offset - prePad);
 			if (result != (ssize_t)blockSize) {
-				[NSException raise:NSGenericException format:@"Read result: %d expected: %u error: %s", result, blockSize, strerror(errno)];
+				[NSException raise:NSGenericException format:@"Read result: %zd expected: %u error: %s", result, blockSize, strerror(errno)];
 			}
 			NSUInteger toCopy = blockSize - prePad;
 			if (toCopy > length)
@@ -363,7 +363,7 @@ static BOOL returnFTruncateError(NSError **error) {
 
 		result = pread(fileDescriptor, tempBuf, blockSize, offset);
 		if (result != (ssize_t)blockSize) {
-			[NSException raise:NSGenericException format:@"Read result: %d expected: %u error: %s", result, blockSize, strerror(errno)];
+			[NSException raise:NSGenericException format:@"Read result: %zd expected: %u error: %s", result, blockSize, strerror(errno)];
 		}
 		memcpy(tempBuf, buff, lastBlockLen);
